Fix bug in compareByQueueMode comparator
The compareByQueueMode is used as a comparison function for sorting.
According to the API documentation on java.lang.Comparable.compareTo,
the implementor must ensure that:

1. sgn(x.compareTo(y)) == -sgn(y.compareTo(x)), for all x and y

2. (x.compareTo(y) > 0 && y.compareTo(z) > 0) implies
   x.compareTo(z) > 0

3. x.compareTo(y) == 0 implies
   sgn(x.compareTo(z)) == sgn(y.compareTo(z)), for all z

The previous implementation of compareByQueueMode violated all three
conditions above. In practice, this can cause problems when sorting,
resulting in a "IllegalArgumentException: Comparison method violates its
general contract".

This patch fixes the compareByQueueMode method such that the three
conditions given above are satisfied.