Fix indentation and use gradle docker image
The .yaml cannot be parsed if the indentation is wrong. Gitlb CI Lint
does not find that and simply complains about the first line.

A gradle docker image is used now. Hopefully that works for us, there is
no official gradle docker image. Alternatively, we could use a java
image, but we would probably need to install gradle on that, so I hope
taking this image makes things easier.