Add missing source folders to gradle script and include sonarqube task/plugin
Up to now, the gradle script only used src/main as source and src/test as testing directory (default settings). I added the missing source folders into the source set. When starting the build, gradle is missing the package "fi.tkk.netlab.dtn.ecla", so the script fails.
I also added the sonarqube task to the build script, so as soon as the build is running, the sonarqube analysis should be working as well.