Check if the host is scanning before even trying new connections
Check for if the host is already connected and just check one interface
I found out we call connect even for all the hosts we were previously
connected to, which will likely not change much within a second. We also
just need to check whether the interface we actually want to connect to
is active, not just any interface of the host. I actually believe that
was a mistake in the code, but never noticed since typical simulations
just include one interface per host.