Add new queue mode to MessageRouter: Highest priority first
In order to test the functionality of our correct message ordering I
need to give messages to the router that should be sent before others
and check whether they are only sent after message reordering. I found
out that since our present modes are random and FIFO (so ordering
depends on when the router got the message), that is not possible.
So introduced a new queue mode, priority. We could use it for evaluation