Allow to configure whether or not connections are bidirectional
Previously, only bidirectional connections were supported. This was
achieved by adding them to both hosts at the same time when the
NetworkInterface connected. In some cases, we have fully-independent up-
and downlink channels (or only one of them), which cannot be
simulated accurately with the default behavior.

This adds a configuration parameter "bidirectionalConnections" on a
per-interface basis, which allows to enable unidirectional connections
(by setting the parameter to false).

This also removes the check from ActiveRouter.checkReceiving whether the
current router is "transferring". This check is redundant because the
outgoing connection is already checked beforehand.